+++ config/routes.rb
@@ -126,17 +126,14 @@ Rails.application.routes.draw do
 
       resources :offers, except: :destroy
     end
-    # show and edit redirect to the nested action:
-    resources :offers, only: [] do
-      collection do
-        get :review
-        post :review_all
-      end
-    end
+
     # show and edit redirect to the nested action:
     resources :offers, only: [:show, :edit, :index] do
+      collection do
+        get :review
+      end
       member do
-        patch :kill
+        patch :kill, :verify
       end
     end
     resources :destinations, only: :index