+++ spec/features/admin/offer_pages_spec.rb
@@ -233,16 +233,17 @@ describe "admin section" do
       describe "when page loads" do
         it "shows only live offers" do
           expect(page).to have_selector( ".offer", count: Offer.live.count)
-          expect(page).to have_button('Done')
         end
       end
 
       describe "when viewing non-reviewed offers" do
         it "shows offer details" do
           is_expected.to have_content @live_1.card.name
-          is_expected.to have_content @live_1.card.bp
+          is_expected.to have_content @live_1.card.bp.to_s[0].upcase
           expect(find("tr#offer_#{@live_1.id}").text).to include('never')
           is_expected.to have_link('Link', href: @live_1.link)
+          is_expected.to have_link "kill_offer_#{ @live_1.id }_btn"
+          is_expected.to have_link "verify_offer_#{ @live_1.id }_btn"
         end
       end
 
@@ -252,27 +253,24 @@ describe "admin section" do
         end
       end
 
-      describe "when pressing Done" do
-        it "updates live offers last_reviewed_at datetime" do
-          click_button("done_btn")
+      describe "when pressing Verify" do
+        it "updates selected last_reviewed_at datetime", :js => true do
+          click_link("verify_offer_#{ @live_1.id }_btn")
+          wait_for_ajax
           @live_1.reload
           expect(@live_1.last_reviewed_at).to be_within(2.seconds).of(Time.now)
+          expect(find("#reviewed_#{@live_1.id}").text).to include(Time.now.strftime("%m/%d/%Y"))
         end
       end
 
-      describe "when pressing Done" do
-        it "does not update dead offers last_reviewed_at datetime"do
+      describe "when pressing Verify" do
+        it "does not update other last_reviewed_at datetimes", :js => true do
           expect do
-            click_button("done_btn")
-            end.not_to change{@dead_1.last_reviewed_at}
-          end
-      end
-
-      describe "when pressing Done" do
-        it "redirects to offers review and alerts success" do
-          click_button("done_btn")
-          expect(page.current_path).to eq review_admin_offers_path
-          is_expected.to have_content "All live offers reviewed"
+            click_link("verify_offer_#{ @live_1.id }_btn")
+            wait_for_ajax
+            @live_2.reload
+            @dead_1.reload
+          end.not_to change{ @live_2.last_reviewed_at }
         end
       end
 
@@ -280,7 +278,7 @@ describe "admin section" do
         it "doesn't kill the offer", :js => true do
           expect do
             page.dismiss_confirm do
-              click_button("kill_offer_#{ @live_1.id }_btn")
+              click_link("kill_offer_#{ @live_1.id }_btn")
             end
           end.not_to change{Offer.live.count}
         end
@@ -289,7 +287,7 @@ describe "admin section" do
       describe "pressing Kill then confirm" do
         it "removes offer from the user display", :js => true do
           page.accept_confirm do
-            find_button("kill_offer_#{ @live_1.id }_btn").click
+            find_link("kill_offer_#{ @live_1.id }_btn").click
           end
           expect(page).to have_selector( ".offer", count: Offer.live.count)
         end
@@ -298,7 +296,7 @@ describe "admin section" do
       describe "pressing Kill then confirm", :js => true do
         it "changes offer live value to false" do
           page.accept_confirm do
-            find_button("kill_offer_#{ @live_2.id }_btn").click
+            find_link("kill_offer_#{ @live_2.id }_btn").click
           end
           wait_for_ajax
           @live_2.reload
@@ -310,7 +308,7 @@ describe "admin section" do
       describe "when killing offers" do
         it "doesnt't delete offers from the database", :js => true do
           page.accept_confirm do
-            find_button("kill_offer_#{ @live_3.id }_btn").click
+            find_link("kill_offer_#{ @live_3.id }_btn").click
           end
           expect(@live_3).to_not be_nil
         end